+++ cfdm/mixin/__init__.py
@@ -8,7 +8,8 @@ from .parameters                  import Parameters
 from .parametersdomainancillaries import ParametersDomainAncillaries
 
 from .netcdf                      import (NetCDFGlobalAttributes,
-                                          NetCDFUnlimitedDimensions,
+#                                          NetCDFUnlimitedDimensions,
+                                          NetCDFUnlimitedDimension,
                                           NetCDFDimension,
                                           NetCDFExternal,
                                           NetCDFGeometry,