+++ cfdm/read_write/netcdf/netcdfread.py
@@ -1971,7 +1971,7 @@ variable should be pre-filled with missing values.
         # ----------------------------------------------------------------
         field_ncdimensions = self._ncdimensions(field_ncvar)
 
-        unlimited = []
+#        unlimited = []
         
         for ncdim in field_ncdimensions:
             if g['variable_dimensions'].get(ncdim) == (ncdim,):
@@ -2004,7 +2004,8 @@ variable should be pre-filled with missing values.
                     
                 # Set unlimited status of axis
                 if nc.dimensions[ncdim].isunlimited():
-                    unlimited.append(axis)
+                    self.implementation.nc_set_unlimited_axis(f, axis)
+#                    unlimited.append(axis)
     
                 ncvar_to_key[ncdim] = dim
                 g['coordinates'].setdefault(field_ncvar, []).append(ncdim)
@@ -2027,7 +2028,8 @@ variable should be pre-filled with missing values.
                 # Set unlimited status of axis
                 try:
                     if nc.dimensions[ncdim].isunlimited():
-                        unlimited.append(axis)
+                        self.implementation.nc_set_unlimited_axis(f, axis)
+#                        unlimited.append(axis)
                 except KeyError:
                     # This dimension is not in the netCDF file (as might
                     # be the case for an element dimension implied by a
@@ -2041,8 +2043,8 @@ variable should be pre-filled with missing values.
             ncdim_to_axis[ncdim] = axis
         #--- End: for
 
-        if unlimited:
-            self.implementation.nc_set_unlimited_dimensions(f, unlimited)
+#        if unlimited:
+#            self.implementation.nc_set_unlimited_dimensions(f, unlimited)
     
         data = self._create_data(field_ncvar, f, unpacked_dtype=unpacked_dtype)