+++ docs/source/releases.rst
@@ -11,6 +11,7 @@
 
 ----
 
+* `Version 1.7.4 <https://ncas-cms.github.io/cfdm/1.7.4>`_
 * `Version 1.7.3 <https://ncas-cms.github.io/cfdm/1.7.3>`_
 * `Version 1.7.2 <https://ncas-cms.github.io/cfdm/1.7.2>`_
 * `Version 1.7.1 <https://ncas-cms.github.io/cfdm/1.7.1>`_